+++ lib/searchkick/index.rb
@@ -38,7 +38,22 @@ module Searchkick
       client.indices.get_settings index: name
     end
 
-    def promote(new_name)
+    def refresh_interval
+      settings.values.first["settings"]["index"]["refresh_interval"]
+    end
+
+    def update_settings(settings)
+      client.indices.put_settings index: name, body: settings
+    end
+
+    def promote(new_name, update_refresh_interval: false)
+      if update_refresh_interval
+        new_index = Searchkick::Index.new(new_name)
+        settings = options[:settings] || {}
+        refresh_interval = (settings[:index] && settings[:index][:refresh_interval]) || "1s"
+        new_index.update_settings(index: {refresh_interval: refresh_interval})
+      end
+
       old_indices =
         begin
           client.indices.get_alias(name: name).keys
@@ -51,26 +66,31 @@ module Searchkick
     alias_method :swap, :promote
 
     # record based
+    # use helpers for notifications
 
     def store(record)
-      bulk_index([record])
+      bulk_index_helper([record])
     end
 
     def remove(record)
-      bulk_delete([record])
+      bulk_delete_helper([record])
+    end
+
+    def update_record(record, method_name)
+      bulk_update_helper([record], method_name)
     end
 
     def bulk_delete(records)
-      Searchkick.indexer.queue(records.reject { |r| r.id.blank? }.map { |r| {delete: record_data(r)} })
+      bulk_delete_helper(records)
     end
 
     def bulk_index(records)
-      Searchkick.indexer.queue(records.map { |r| {index: record_data(r).merge(data: search_data(r))} })
+      bulk_index_helper(records)
     end
     alias_method :import, :bulk_index
 
     def bulk_update(records, method_name)
-      Searchkick.indexer.queue(records.map { |r| {update: record_data(r).merge(data: {doc: search_data(r, method_name)})} })
+      bulk_update_helper(records, method_name)
     end
 
     def record_data(r)
@@ -131,8 +151,15 @@ module Searchkick
       search_model(record.class, like_text, options)
     end
 
+    # queue
+
+    def reindex_queue
+      Searchkick::ReindexQueue.new(name)
+    end
+
     # search
 
+    # TODO remove in next major version
     def search_model(searchkick_klass, term = "*", **options, &block)
       query = Searchkick::Query.new(searchkick_klass, term, options)
       yield(query.body) if block
@@ -187,44 +214,55 @@ module Searchkick
 
     # https://gist.github.com/jarosan/3124884
     # http://www.elasticsearch.org/blog/changing-mapping-with-zero-downtime/
-    def reindex_scope(scope, import: true, resume: false)
+    def reindex_scope(scope, import: true, resume: false, retain: false, async: false, refresh_interval: nil)
       if resume
         index_name = all_indices.sort.last
         raise Searchkick::Error, "No index to resume" unless index_name
         index = Searchkick::Index.new(index_name)
       else
-        clean_indices
+        clean_indices unless retain
 
-        index = create_index(index_options: scope.searchkick_index_options)
+        index_options = scope.searchkick_index_options
+        index_options.deep_merge!(settings: {index: {refresh_interval: refresh_interval}}) if refresh_interval
+        index = create_index(index_options: index_options)
       end
 
       # check if alias exists
       if alias_exists?
         # import before promotion
-        index.import_scope(scope, resume: resume) if import
+        index.import_scope(scope, resume: resume, async: async, full: true) if import
 
         # get existing indices to remove
-        promote(index.name)
-        clean_indices
+        unless async
+          promote(index.name, update_refresh_interval: !refresh_interval.nil?)
+          clean_indices unless retain
+        end
       else
         delete if exists?
-        promote(index.name)
+        promote(index.name, update_refresh_interval: !refresh_interval.nil?)
 
         # import after promotion
-        index.import_scope(scope, resume: resume) if import
+        index.import_scope(scope, resume: resume, async: async, full: true) if import
       end
 
-      index.refresh
-
-      true
+      if async
+        {index_name: index.name}
+      else
+        index.refresh
+        true
+      end
     end
 
-    def import_scope(scope, resume: false, method_name: nil)
-      batch_size = @options[:batch_size] || 1000
-
+    def import_scope(scope, resume: false, method_name: nil, async: false, batch: false, batch_id: nil, full: false)
       # use scope for import
       scope = scope.search_import if scope.respond_to?(:search_import)
-      if scope.respond_to?(:find_in_batches)
+
+      if batch
+        import_or_update scope.to_a, method_name, async
+        Searchkick.with_redis { |r| r.srem(batches_key, batch_id) } if batch_id
+      elsif full && async
+        full_reindex_async(scope)
+      elsif scope.respond_to?(:find_in_batches)
         if resume
           # use total docs instead of max id since there's not a great way
           # to get the max _id without scripting since it's a string
@@ -233,36 +271,20 @@ module Searchkick
           scope = scope.where("id > ?", total_docs)
         end
 
+        scope = scope.select("id").except(:includes, :preload) if async
+
         scope.find_in_batches batch_size: batch_size do |batch|
-          import_or_update batch.select(&:should_index?), method_name
+          import_or_update batch, method_name, async
         end
       else
-        # https://github.com/karmi/tire/blob/master/lib/tire/model/import.rb
-        # use cursor for Mongoid
-        items = []
-        # TODO add resume
-        scope.all.each do |item|
-          items << item if item.should_index?
-          if items.length == batch_size
-            import_or_update items, method_name
-            items = []
-          end
+        each_batch(scope) do |items|
+          import_or_update items, method_name, async
         end
-        import_or_update items, method_name
       end
     end
 
-    def import_or_update(records, method_name)
-      retries = 0
-      begin
-        method_name ? bulk_update(records, method_name) : import(records)
-      rescue Faraday::ClientError => e
-        if retries < 1
-          retries += 1
-          retry
-        end
-        raise e
-      end
+    def batches_left
+      Searchkick.with_redis { |r| r.scard(batches_key) }
     end
 
     # other
@@ -372,6 +394,107 @@ module Searchkick
       else
         obj
       end
+    end
+
+    def import_or_update(records, method_name, async)
+      if records.any?
+        if async
+          Searchkick::BulkReindexJob.perform_later(
+            class_name: records.first.class.name,
+            record_ids: records.map(&:id),
+            index_name: name,
+            method_name: method_name ? method_name.to_s : nil
+          )
+        else
+          records = records.select(&:should_index?)
+          if records.any?
+            with_retries do
+              method_name ? bulk_update(records, method_name) : import(records)
+            end
+          end
+        end
+      end
+    end
+
+    def full_reindex_async(scope)
+      if scope.respond_to?(:primary_key)
+        # TODO expire Redis key
+        primary_key = scope.primary_key
+        starting_id = scope.minimum(primary_key) || 0
+        max_id = scope.maximum(primary_key) || 0
+        batches_count = ((max_id - starting_id + 1) / batch_size.to_f).ceil
+
+        batches_count.times do |i|
+          batch_id = i + 1
+          min_id = starting_id + (i * batch_size)
+          bulk_reindex_job scope, batch_id, min_id: min_id, max_id: min_id + batch_size - 1
+        end
+      else
+        batch_id = 1
+        # TODO remove any eager loading
+        scope = scope.only(:_id) if scope.respond_to?(:only)
+        each_batch(scope) do |items|
+          bulk_reindex_job scope, batch_id, record_ids: items.map { |i| i.id.to_s }
+          batch_id += 1
+        end
+      end
+    end
+
+    def each_batch(scope)
+      # https://github.com/karmi/tire/blob/master/lib/tire/model/import.rb
+      # use cursor for Mongoid
+      items = []
+      scope.all.each do |item|
+        items << item
+        if items.length == batch_size
+          yield items
+          items = []
+        end
+      end
+      yield items if items.any?
+    end
+
+    def bulk_reindex_job(scope, batch_id, options)
+      Searchkick::BulkReindexJob.perform_later({
+        class_name: scope.model_name.name,
+        index_name: name,
+        batch_id: batch_id
+      }.merge(options))
+      Searchkick.with_redis { |r| r.sadd(batches_key, batch_id) }
+    end
+
+    def batch_size
+      @batch_size ||= @options[:batch_size] || 1000
+    end
+
+    def with_retries
+      retries = 0
+
+      begin
+        yield
+      rescue Faraday::ClientError => e
+        if retries < 1
+          retries += 1
+          retry
+        end
+        raise e
+      end
+    end
+
+    def bulk_index_helper(records)
+      Searchkick.indexer.queue(records.map { |r| {index: record_data(r).merge(data: search_data(r))} })
+    end
+
+    def bulk_delete_helper(records)
+      Searchkick.indexer.queue(records.reject { |r| r.id.blank? }.map { |r| {delete: record_data(r)} })
+    end
+
+    def bulk_update_helper(records, method_name)
+      Searchkick.indexer.queue(records.map { |r| {update: record_data(r).merge(data: {doc: search_data(r, method_name)})} })
+    end
+
+    def batches_key
+      "searchkick:reindex:#{name}:batches"
     end
   end
 end