+++ lib/searchkick/results.rb
@@ -184,6 +184,10 @@ module Searchkick
       @response["hits"]["hits"]
     end
 
+    def misspellings?
+      @options[:misspellings]
+    end
+
     private
 
     def results_query(records, hits)
@@ -202,21 +206,7 @@ module Searchkick
           end
       end
 
-      records =
-        if records.respond_to?(:primary_key)
-          # ActiveRecord
-          records.where(records.primary_key => ids) if records.primary_key
-        elsif records.respond_to?(:queryable)
-          # Mongoid 3+
-          records.queryable.for_ids(ids)
-        elsif records.respond_to?(:unscoped) && :id.respond_to?(:in)
-          # Nobrainer
-          records.unscoped.where(:id.in => ids)
-        end
-
-      raise Searchkick::Error, "Not sure how to load records" if !records
-
-      records
+      Searchkick.load_records(records, ids)
     end
 
     def base_field(k)