+++ test/test_helper.rb
@@ -14,12 +14,24 @@ File.delete("elasticsearch.log") if File.exist?("elasticsearch.log")
 Searchkick.client.transport.logger = Logger.new("elasticsearch.log")
 Searchkick.search_timeout = 5
 
+if defined?(Redis)
+  if defined?(ConnectionPool)
+    Searchkick.redis = ConnectionPool.new { Redis.new }
+  else
+    Searchkick.redis = Redis.new
+  end
+end
+
 puts "Running against Elasticsearch #{Searchkick.server_version}"
 
 I18n.config.enforce_available_locales = true
 
-ActiveJob::Base.logger = nil if defined?(ActiveJob)
-ActiveSupport::LogSubscriber.logger = Logger.new(STDOUT) if ENV["NOTIFICATIONS"]
+if defined?(ActiveJob)
+  ActiveJob::Base.logger = nil
+  ActiveJob::Base.queue_adapter = :inline
+end
+
+ActiveSupport::LogSubscriber.logger = ActiveSupport::Logger.new(STDOUT) if ENV["NOTIFICATIONS"]
 
 def elasticsearch_below50?
   Searchkick.server_below?("5.0.0-alpha1")
@@ -31,6 +43,10 @@ end
 
 def nobrainer?
   defined?(NoBrainer)
+end
+
+def cequel?
+  defined?(Cequel)
 end
 
 if defined?(Mongoid)
@@ -167,6 +183,88 @@ elsif defined?(NoBrainer)
 
   class Cat < Animal
   end
+elsif defined?(Cequel)
+  cequel =
+    Cequel.connect(
+      host: "127.0.0.1",
+      port: 9042,
+      keyspace: "searchkick_test",
+      default_consistency: :all
+    )
+  # cequel.logger = ActiveSupport::Logger.new(STDOUT)
+  cequel.schema.drop! if cequel.schema.exists?
+  cequel.schema.create!
+  Cequel::Record.connection = cequel
+
+  class Product
+    include Cequel::Record
+
+    key :id, :uuid, auto: true
+    column :name, :text, index: true
+    column :store_id, :int
+    column :in_stock, :boolean
+    column :backordered, :boolean
+    column :orders_count, :int
+    column :found_rate, :decimal
+    column :price, :int
+    column :color, :text
+    column :latitude, :decimal
+    column :longitude, :decimal
+    column :description, :text
+    column :alt_description, :text
+    column :created_at, :timestamp
+  end
+
+  class Store
+    include Cequel::Record
+
+    key :id, :timeuuid, auto: true
+    column :name, :text
+
+    # has issue with id serialization
+    def search_data
+      {
+        name: name
+      }
+    end
+  end
+
+  class Region
+    include Cequel::Record
+
+    key :id, :timeuuid, auto: true
+    column :name, :text
+    column :text, :text
+  end
+
+  class Speaker
+    include Cequel::Record
+
+    key :id, :timeuuid, auto: true
+    column :name, :text
+  end
+
+  class Animal
+    include Cequel::Record
+
+    key :id, :timeuuid, auto: true
+    column :name, :text
+
+    # has issue with id serialization
+    def search_data
+      {
+        name: name
+      }
+    end
+  end
+
+  class Dog < Animal
+  end
+
+  class Cat < Animal
+  end
+
+  [Product, Store, Region, Speaker, Animal].each(&:synchronize_schema)
 else
   require "active_record"
 
@@ -418,7 +516,8 @@ class Animal
   searchkick \
     text_start: [:name],
     suggest: [:name],
-    index_name: -> { "#{name.tableize}-#{Date.today.year}" }
+    index_name: -> { "#{name.tableize}-#{Date.today.year}" },
+    callbacks: defined?(ActiveJob) ? :async : true
     # wordnet: true
 end